+++ src/net-util/export/UriUtil.js
@@ -11,46 +11,6 @@ import { MustBe } from '@this/typey';
  */
 export class UriUtil {
   /**
-   * Checks that a given value is a string which can be interpreted as an
-   * absolute URI path (no protocol, host, etc.). It must:
-   *
-   * * Start with a slash (`/`).
-   * * End with a slash (`/`).
-   * * Not contain double (or more) slashes.
-   * * Not contain `.` or `..` path components.
-   * * Not contain a query or hash fragment.
-   * * Not contain characters that need `%`-encoding. (It is expected to be
-   *   pre-encoded.)
-   *
-   * @param {*} value Value in question.
-   * @returns {string} `value` if it is a string which matches the pattern.
-   * @throws {Error} Thrown if `value` does not match.
-   */
-  static checkAbsolutePath(value) {
-    // Basic constraints.
-    const pattern =
-      '^' +
-      '(?=[/])' +              // Must start with a slash.
-      '(?!.*[/][.]{0,2}[/])' + // No empty, `.`, or `..` components.
-      '.*/$';                  // Must end with a slash.
-
-    MustBe.string(value, pattern);
-
-    // Check the rest of the constraints by parsing and only accepting it if
-    // a successfully-parsed path is the same as the given one.
-    try {
-      const url = new URL(`http://x${value}`);
-      if (url.pathname === value) {
-        return value;
-      }
-    } catch {
-      // Fall through to throw the error.
-    }
-
-    throw new Error('Must be an absolute URI path.');
-  }
-
-  /**
    * Checks that a given value is a string which can be interpreted as a "basic"
    * absolute URI. It must:
    *
@@ -95,17 +55,36 @@ export class UriUtil {
   }
 
   /**
-   * Checks that a given value is a string representing a protocol name (as
-   * allowed by this system).
+   * Checks if a given string is syntactically valid as a received URI path
+   * component (thing between slashes or at the end of a path before queries,
+   * etc.), per RFC 3986. The caveat "received" is that `.` and `..` are
+   * supposed to get "resolved away" before being sent to a server.
    *
-   * @param {*} value Value in question.
-   * @returns {string} `value` if it is a string which matches the stated
-   *   pattern.
-   * @throws {Error} Thrown if `value` does not match.
+   * @param {*} component Alleged path component.
+   * @returns {boolean} `true` if `component` is actually syntactically valid.
    */
-  static checkProtocol(value) {
-    const pattern = /^(http|https|http2)$/;
-    return MustBe.string(value, pattern);
+  static isPathComponent(component) {
+    MustBe.string(component);
+
+    if ((component === '.') || (component === '..')) {
+      return false;
+    }
+
+    // Allow ASCII alphanumerics, plus any of `-_.~!$&'()*+,;=:@%`.
+    if (!/^[-_.~!$&'()*+,;=:@%A-Za-z0-9]*$/.test(component)) {
+      return false;
+    }
+
+    // If there are any `%`s, they must be valid percent-encoded sequences.
+    if (/%/.test(component)) {
+      for (const [match] of component.matchAll(/%.{0,2}/g)) {
+        if (!/^%[0-9A-F]{2}$/.test(match)) {
+          return false;
+        }
+      }
+    }
+
+    return true;
   }
 
   /**